backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-socketio==5.10.0
pydantic==2.5.2
pydantic-settings==2.1.0
python-dotenv==1.0.0

# OpenAI
openai>=1.0.0

# Vector DB
chromadb>=0.4.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Redis
redis>=5.0.0

# File handling
python-multipart>=0.0.6
aiofiles>=23.0.0
